+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/WebViewServer.kt
@@ -18,6 +18,7 @@ import org.readium.r2.navigator.epub.css.ReadiumCss
 import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Publication
+import org.readium.r2.shared.publication.UrlHref
 import org.readium.r2.shared.resource.Resource
 import org.readium.r2.shared.resource.ResourceInputStream
 import org.readium.r2.shared.resource.StringResource
@@ -25,7 +26,6 @@ import org.readium.r2.shared.resource.fallback
 import org.readium.r2.shared.util.AbsoluteUrl
 import org.readium.r2.shared.util.Try
 import org.readium.r2.shared.util.Url
-import org.readium.r2.shared.util.UrlHref
 import org.readium.r2.shared.util.http.HttpHeaders
 import org.readium.r2.shared.util.http.HttpRange
 import org.readium.r2.shared.util.mediatype.MediaType
