+++ readium/shared/src/main/java/org/readium/r2/shared/publication/Href.kt
@@ -4,10 +4,12 @@
  * available in the top-level LICENSE file of the project.
  */
 
-package org.readium.r2.shared.util
+package org.readium.r2.shared.publication
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import org.readium.r2.shared.util.URITemplate
+import org.readium.r2.shared.util.Url
 
 public sealed class Href : Parcelable {
 
@@ -35,4 +37,18 @@ public data class TemplatedHref(val template: String) : Href() {
     override fun toUrl(base: Url?, parameters: Map<String, String>): Url? =
         Url(URITemplate(template).expand(parameters))
             ?.let { base?.resolve(it) ?: it }
+}
+
+/**
+ * Resolves the given [Href] to this URL.
+ */
+internal fun Url?.resolve(href: Href): Href {
+    this ?: return href
+
+    return when (href) {
+        is TemplatedHref -> Url(href.template)
+            ?.let { TemplatedHref(resolve(it).toString()) }
+            ?: href
+        is UrlHref -> UrlHref(resolve(href.url))
+    }
 }
