+++ readium/streamer/src/main/java/org/readium/r2/streamer/parser/epub/NcxParser.kt
@@ -8,8 +8,8 @@ package org.readium.r2.streamer.parser.epub
 
 import org.readium.r2.shared.parser.xml.ElementNode
 import org.readium.r2.shared.publication.Link
+import org.readium.r2.shared.publication.UrlHref
 import org.readium.r2.shared.util.Url
-import org.readium.r2.shared.util.UrlHref
 
 internal object NcxParser {
