+++ readium/streamer/src/test/java/org/readium/r2/streamer/TestUtils.kt
@@ -10,11 +10,11 @@
 package org.readium.r2.streamer
 
 import kotlinx.coroutines.runBlocking
+import org.readium.r2.shared.publication.Href
 import org.readium.r2.shared.publication.Publication
+import org.readium.r2.shared.publication.UrlHref
 import org.readium.r2.shared.resource.Resource
-import org.readium.r2.shared.util.Href
 import org.readium.r2.shared.util.Url
-import org.readium.r2.shared.util.UrlHref
 import org.readium.r2.streamer.parser.PublicationParser
 
 internal fun Resource.readBlocking(range: LongRange? = null) = runBlocking { read(range) }
